% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControl.R
\name{selectControlInput}
\alias{selectControlInput}
\title{<Add Title>}
\usage{
selectControlInput(
  inputId,
  label,
  styles,
  choices,
  selected = NULL,
  multiple = FALSE,
  containerClass = "mt-4 col-md-6 col-offset-4",
  animated = FALSE,
  displayGroupSizes = TRUE
)
}
\description{
<Add Description>
}
