% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControl.R
\name{toggleMenu}
\alias{toggleMenu}
\title{Toggle a select control widget}
\usage{
toggleMenu(session, inputId)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{inputId}{the input id of the select control}
}
\value{
No value; called for side effect.
}
\description{
Toggle (open/close) a select control widget.
}
\examples{
# See the last example of 'selectControlInput'.
}
