% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControl.R
\name{updateSelectControlInput}
\alias{updateSelectControlInput}
\title{Update a select control widget}
\usage{
updateSelectControlInput(session, inputId, choices = NULL, selected = NULL)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{inputId}{the id of the select control widget to be updated}

\item{choices}{new choices, or \code{NULL}}

\item{selected}{new value(s) for the selected items, or \code{NULL}}
}
\value{
No returned value, called for side effect.
}
\description{
Change the value of a select control input.
}
